parser: AST api (#17)
This MR brings the following changes.

At this point the AST nodes representing either type or operation are wired through the MathPotatoAstTree and return a tuple with Id and type. This MR formulates a decent Api for the MathPotatoAstTree when it comes to manage the nodes in the tree (actually tree per type). As a consequence every type has an internal type (this one is without an id, as the id is the key of the hashmap these types stored) and a customer facing type. The latter includes the id.